added category_id to product

+++ database/migrations/2020_03_14_205049_create_group_products_table.php
@@ -17,7 +17,7 @@ class CreateGroupProductsTable extends Migration
             $table->id();
             $table->bigInteger('product_id')->unsigned();
             $table->bigInteger('group_id')->unsigned();
-            $table->tinyInteger('order_no');
+            $table->integer('order_no');
             $table->bigInteger('updated_by')->nullable();
             $table->bigInteger('created_by');
             $table->timestamps();
