+++ database/migrations/2020_03_14_220325_create_group_categories_table.php
@@ -17,7 +17,7 @@ class CreateGroupCategoriesTable extends Migration
             $table->id();
             $table->bigInteger('category_id')->unsigned();
             $table->bigInteger('group_id')->unsigned();
-            $table->tinyInteger('order_no');
+            $table->integer('order_no');
             $table->bigInteger('updated_by')->nullable();
             $table->bigInteger('created_by');
             $table->timestamps();
